# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /apartment-pairs:
    get:
      operationId: getApartmentPairs
      tags: [apartment-pairs]
      summary: Get ranked Alicante apartment pairs
      description: Returns source-linked apartment pairs with explicit verification and availability status.
      parameters:
        - name: status
          in: query
          required: false
          description: Limit results to source verification status.
          schema:
            type: string
            enum: [all, verified, provisional]
            default: all
        - name: area
          in: query
          required: false
          description: Limit results to pairs that include this Alicante area.
          schema:
            type: string
        - name: sort
          in: query
          required: false
          description: Ranking order for the returned pairs.
          schema:
            type: string
            enum: [rank, date-found, beach-distance]
            default: rank
      responses:
        "200":
          description: Ranked apartment pairs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApartmentPairsResponse"
  /apartment-pairs/import:
    post:
      operationId: importApartmentPairPublication
      tags: [apartment-pairs]
      summary: Import a validated apartment-pair publication
      description: Atomically merges into the active publication. Omitted pairs remain visible; matching apartment compositions update under their existing pair ID. Explicit reasoned archive requests remove pairs, and archived pairs require explicit restoration. Requires the private importer bearer token.
      security:
        - apartmentImporterToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApartmentPairPublicationInput"
      responses:
        "201":
          description: Publication imported and activated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApartmentPairImportResult"
        "400":
          description: Invalid publication
        "401":
          description: Invalid or missing importer token
        "503":
          description: Importer token is not configured
components:
  securitySchemes:
    apartmentImporterToken:
      type: http
      scheme: bearer
      bearerFormat: opaque
  schemas:
...
    ApartmentPhoto:
      type: object
      properties:
        url:
          type: string
          format: uri
        alt:
          type: string
        sourceUrl:
          type: string
          format: uri
        attribution:
          type: string
        aiEdited:
          type: boolean
      required: [url, alt, sourceUrl, attribution, aiEdited]
    ApartmentPairListing:
      type: object
      properties:
        id:
          type: string
          description: Stable source record ID.
        displayName:
          type: string
        area:
          type: string
        address:
          type: ["string", "null"]
        bedrooms:
          type: ["integer", "null"]
        sizeSqm:
          type: ["number", "null"]
        monthlyRentEur:
          type: ["number", "null"]
        furnished:
          type: ["boolean", "null"]
        availabilityStatus:
          type: string
          enum: [confirmed, unknown, unavailable]
        availabilityNote:
          type: string
        originalListingUrl:
          type: string
          format: uri
        photoSourceUrl:
          type: ["string", "null"]
          format: uri
        photoAttribution:
          type: ["string", "null"]
        photos:
          type: array
          items:
            $ref: "#/components/schemas/ApartmentPhoto"
        sourceName:
          type: string
        sourceCheckedAt:
          type: string
      required:
        - id
        - displayName
        - area
        - address
        - bedrooms
        - sizeSqm
        - monthlyRentEur
        - furnished
        - availabilityStatus
        - availabilityNote
        - originalListingUrl
        - photoSourceUrl
        - photoAttribution
        - photos
        - sourceName
        - sourceCheckedAt
    ApartmentPair:
      type: object
      properties:
        id:
          type: string
          description: Stable pair ID.
        rank:
          type: integer
        paperId:
          type: string
          description: Stable Paper record ID for the pair.
        apartmentId:
          type: string
          description: Stable Apartment record ID for the pair.
        title:
          type: string
        summary:
          type: string
        status:
          type: string
          enum: [verified, provisional]
        statusNote:
          type: string
        areas:
          type: array
          items:
            type: string
        beachDistanceMinutes:
          type: ["integer", "null"]
        combinedMonthlyRentEur:
          type: ["number", "null"]
        poolDistance:
          type: string
        beachDistance:
          type: string
        groceriesDistance:
          type: string
        cafesDistance:
          type: string
        pairWalkingDistance:
          type: string
        pairWalkingTime:
          type: string
        paper:
          $ref: "#/components/schemas/ApartmentPairListing"
        apartment:
          $ref: "#/components/schemas/ApartmentPairListing"
      required:
        - id
        - rank
        - paperId
        - apartmentId
        - title
        - summary
        - status
        - statusNote
        - areas
        - beachDistanceMinutes
        - combinedMonthlyRentEur
        - poolDistance
        - beachDistance
        - groceriesDistance
        - cafesDistance
        - pairWalkingDistance
        - pairWalkingTime
        - paper
        - apartment
    ApartmentPairsResponse:
      type: object
      properties:
        reviewedAt:
          type: string
        disclaimer:
          type: string
        total:
          type: integer
        pairs:
          type: array
          items:
            $ref: "#/components/schemas/ApartmentPair"
        availableApartments:
          type: array
          description: Independently valid apartments retained for possible future pairing.
          items:
            $ref: "#/components/schemas/ApartmentPairListing"
      required: [reviewedAt, disclaimer, total, pairs, availableApartments]
    ApartmentPairBudgetException:
      type: object
      properties:
        pairId:
          type: string
          minLength: 1
        reason:
          type: string
          minLength: 20
      required: [pairId, reason]
    ApartmentPairPublicationInput:
      type: object
      properties:
        reviewedAt:
          type: string
          minLength: 1
        disclaimer:
          type: string
          minLength: 1
        pairs:
          type: array
          minItems: 0
          items:
            $ref: "#/components/schemas/ApartmentPair"
        archivePairs:
          description: Explicit removals by existing pair ID. Omission never archives a pair. Reasons are retained with publication history.
          type: array
          items:
            type: object
            properties:
              pairId:
                type: string
                minLength: 1
              reason:
                type: string
                minLength: 20
            required: [pairId, reason]
        restorePairIds:
          description: Explicitly restore archived pair IDs; include a freshly checked complete pair with the same apartment composition in pairs.
          type: array
          items:
            type: string
            minLength: 1
        budgetExceptions:
          type: array
          items:
            $ref: "#/components/schemas/ApartmentPairBudgetException"
          default: []
      required: [reviewedAt, disclaimer, pairs, budgetExceptions]
    ApartmentPairImportResult:
      type: object
      properties:
        publicationId:
          type: string
          format: uuid
        importedAt:
          type: string
          format: date-time
        total:
          type: integer
      required: [publicationId, importedAt, total]